added more tests and fixed path catenation

+++ include/fileos/config.h
@@ -68,13 +68,13 @@ IN THE SOFTWARE.
 
 namespace fileos {
 
-typedef signed short int16_t;
-typedef unsigned short uint16_t;
-typedef signed int int32_t;
-typedef unsigned int uint32_t;
-typedef signed long long int64_t;
-typedef unsigned long long uint64_t;
-typedef char utf8_t;
+typedef signed short        int16_t;
+typedef signed int          int32_t;
+typedef signed long long    int64_t;
+typedef unsigned short      uint16_t;
+typedef unsigned int        uint32_t;
+typedef unsigned long long  uint64_t;
+typedef char                utf8_t;
 
 } // end of fileos
